Check if column exists before removing it

+++ db/migrate/20220224173405_remove_original_filename_from_documents.rb
@@ -1,5 +1,5 @@
 class RemoveOriginalFilenameFromDocuments < ActiveRecord::Migration[5.1]
   def change
-    remove_column :documents, :original_filename
+    remove_column :documents, :original_filename if column_exists?(:documents, :original_filename)
   end
 end
